feat: move ListPlugin to seperate folder
