+++ packages/runtime/slate.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/slate.cjs')
