+++ packages/runtime/src/slate/ListPlugin/test/slate-test-helper.tsx
@@ -19,7 +19,7 @@ import {
   Typography,
   Unordered,
 } from './types'
-import { TextType, BlockType } from '../../../../../../controls'
+import { TextType, BlockType } from '../../../controls'
 import { withList } from '../ListPlugin'
 
 export const jsx = createHyperscript({
